+++ exporters/variables-scss/src/content/generators/fileGenerator.ts
@@ -1,51 +1,52 @@
-import { Token, TokenGroup, TokenType } from '@supernovaio/sdk-exporters';
-import { exportConfiguration } from '../../index';
-import { convertTokensToCSS, generateCssObject } from './cssGenerator';
-import { dimensionTokenToCSS } from './dimension';
+import { TokenGroup, Token, TokenType } from '@supernovaio/sdk-exporters';
+import { createFileWithContent } from './contentGenerator';
 
-const addDisclaimer = (content: string): string => {
-  if (exportConfiguration.generateDisclaimer) {
-    return `/* This file was generated by Supernova, don't change manually */\n${content}`;
-  }
+const filesData = [
+  {
+    fileName: '_spaces.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Spacing'],
+    withCssObject: true,
+    withParent: false,
+  },
+  {
+    fileName: '_radii.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Radius'],
+    withCssObject: true,
+    withParent: false,
+  },
+  {
+    fileName: '_borders.scss',
+    tokenTypes: [TokenType.dimension],
+    groupNames: ['Border'],
+    withCssObject: false,
+    withParent: true,
+  },
+  {
+    fileName: '_other.scss',
+    tokenTypes: [TokenType.dimension, TokenType.string],
+    groupNames: ['Grid', 'Container', 'Breakpoint'],
+    withCssObject: true,
+    withParent: true,
+  },
+];
 
-  return content;
-};
-
-export const createFileWithContent = (
-  tokens: Token[],
+export const generateFiles = (
+  tokens: Array<Token>,
   mappedTokens: Map<string, Token>,
   tokenGroups: Array<TokenGroup>,
-  fileName: string,
-  tokenTypes: TokenType[],
-  groupNames: string[],
-  withCssObject: boolean,
-  withParent: boolean = false,
 ) => {
-  let cssTokens = '';
-  let cssObject = '';
-
-  tokenTypes.forEach((tokenType) => {
-    groupNames.forEach((group) => {
-      const filteredTokens = tokens.filter(
-        (token) => token.tokenType === tokenType && token.origin?.name?.includes(group),
-      );
-
-      cssTokens += convertTokensToCSS(filteredTokens, dimensionTokenToCSS, mappedTokens, tokenGroups, withParent);
-      cssTokens += '\n\n';
-
-      const tempCssObject = generateCssObject(filteredTokens, mappedTokens, tokenGroups, withParent);
-      if (tempCssObject !== null) {
-        cssObject += tempCssObject;
-      }
-    });
+  return filesData.map(({ fileName, tokenTypes, groupNames, withCssObject, withParent }) => {
+    return createFileWithContent(
+      tokens,
+      mappedTokens,
+      tokenGroups,
+      fileName,
+      tokenTypes,
+      groupNames,
+      withCssObject,
+      withParent,
+    );
   });
-
-  let content = withCssObject ? `${cssTokens}${cssObject}` : cssTokens;
-  // Remove extra blank lines
-  content = content.replace(/\n{3,}/g, '\n\n');
-
-  return {
-    fileName,
-    content: addDisclaimer(content),
-  };
 };
